change:experiment_data is now an array, not dict
The `experiment_data[type]` collection for elements was changed from dictionary to list since a dictionary neccesitates that the key be known or all values should be traveresed to get the values. The list holds all objects of a given type, with the label lookup handled by the new var `experiment_data_labels`.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>