new:export elements and diagram nodes, import file
The serialize file now contains an export function that flattens the experiment_data objects (no circular or direct object references) and exports it to JSON which is then served as a file through the browser. The import function (stub) extracts and parses the contents of the file from a file picker.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>