new:add multiple text fields to property
Text fields with multiple values now have multiple input fields. They are validated, saved, exported, and imported correctly.

Note: The fields can only be added, and cannot be removed. For adding fields, a '+' button is added to the field label. The corresponding '-' button was not added due to layout issues. The button is added towards the bottom of the label instead of next to the input field.
Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>